More bot logic
This commit adds a bunch of changes to the bot logic such as. Modifying
the bot actions function to be more of how it should look when finished.
What should be the finished logic for when the player has the center
tile. Along with an unfinished test for the logic for when player has
center.